+++ spec/hunspell-asm/mountDirectory-spec.ts
@@ -1,16 +1,17 @@
+//tslint:disable:no-require-imports
 import { expect } from 'chai';
+import { ENVIRONMENT } from 'emscripten-wasm-loader';
 import * as unixify from 'unixify';
 import { FS } from '../../src/HunspellAsmModule';
 
 describe('mountDirectory', () => {
   const nodePathId: string = 'nodePathPrefixDummy';
-  let mountDirectory: (dirPath: string) => string;
   let fsMock: FS;
 
   beforeEach(() => {
     jest.mock('path');
     jest.mock('../../src/isMounted');
-    jest.mock('emscripten-wasm-loader', () => ({ isNode: jest.fn() }));
+    //jest.mock('emscripten-wasm-loader', () => ({ isNode: jest.fn() }));
     jest.mock('../../src/mkdirTree');
 
     fsMock = {
@@ -19,21 +20,16 @@ describe('mountDirectory', () => {
       },
       mount: jest.fn()
     } as any;
-
-    //tslint:disable-next-line:no-require-imports
-    mountDirectory = require('../../src/mountDirectory').mountDirectory(fsMock, nodePathId);
   });
 
   it('should throw if environment is not node', () => {
-    //tslint:disable-next-line:no-require-imports
-    (require('emscripten-wasm-loader').isNode as jest.Mock<any>).mockReturnValueOnce(false);
-
-    expect(() => mountDirectory('/user')).to.throw();
+    const mountDirectory = require('../../src/mountDirectory').mountDirectory(fsMock, nodePathId, ENVIRONMENT.WEB);
+    expect(() => mountDirectory('/user', ENVIRONMENT.WEB)).to.throw();
   });
 
   it('should return if path is already mounted', () => {
     //tslint:disable:no-require-imports
-    (require('emscripten-wasm-loader').isNode as jest.Mock<any>).mockReturnValueOnce(true);
+    const mountDirectory = require('../../src/mountDirectory').mountDirectory(fsMock, nodePathId, ENVIRONMENT.NODE);
     (require('path').join as jest.Mock<any>).mockImplementationOnce((...args: Array<any>) => args.join('/'));
     (require('path').resolve as jest.Mock<any>).mockImplementationOnce((arg: string) => arg);
     (require('../../src/isMounted').isMounted as jest.Mock<any>).mockReturnValueOnce(true);
@@ -49,7 +45,7 @@ describe('mountDirectory', () => {
 
   it('should create and mount for new path provided', () => {
     //tslint:disable:no-require-imports
-    (require('emscripten-wasm-loader').isNode as jest.Mock<any>).mockReturnValueOnce(true);
+    const mountDirectory = require('../../src/mountDirectory').mountDirectory(fsMock, nodePathId, ENVIRONMENT.NODE);
     (require('path').join as jest.Mock<any>).mockImplementationOnce((...args: Array<any>) => args.join('/'));
     (require('path').resolve as jest.Mock<any>).mockImplementation((arg: string) => arg);
     (require('../../src/isMounted').isMounted as jest.Mock<any>).mockReturnValueOnce(false);
@@ -68,3 +64,4 @@ describe('mountDirectory', () => {
     expect(mkdirTreeMock.mock.calls[0]).to.deep.equal([fsMock, expected]);
   });
 });
+//tslint:enable:no-require-imports